code/agents/choice_maker/components/questions_generator/requirements.txt
PyYAML>=6.0
pydantic>=2.7
httpx>=0.26
openai>=1.14
numpy>=1.26
scikit-learn>=1.4
rapidfuzz>=3.6
rich>=13.7
